Modify shared.Field table definition to handle LBM scans
Currently, shared.Field() is created with 25 fields. This is not sufficient to handle the expected range of fields per scan for LBM scans, e.g. 5 fields per 30 depths = 150 fields total. 

Additionally, the datatype used in this table should be changed from tinyint to smallint, as tinyint only has a range of (-128 to 127). smallint has a more-than-sufficient range of (-32,768 to 32,767).

This table will likely not be re-created prior to the development of a/the pipeline. I'm making this PR to remind those in the future to consider this when making design choices for that new pipeline. Feel free to leave this PR un-approved until that time.